+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   implicitly used.
 - `ForbiddenFieldRule` analysis rule template, which flags fields from a specified list of names on
   a type with a specified fully-qualified name.
+- `ForbiddenConstantRule` analysis rule template, which flags constants from a specified list of
+  names from a unit with a specified fully-qualified name.
 
 ### Changed